+++ src/project/message/FindNodeMessage.java
@@ -5,7 +5,7 @@ import java.math.BigInteger;
 public class FindNodeMessage extends BaseMessage{
     private final BigInteger key;
 
-    public FindNodeMessage(Message_Type type, int sender_id, BigInteger key) {
+    public FindNodeMessage(Message_Type type, BigInteger sender_id, BigInteger key) {
         super(type, sender_id);
         this.key = key;
     }
