+++ src/project/message/GetChunkMessage.java
@@ -1,9 +1,11 @@
 package project.message;
 
+import java.math.BigInteger;
+
 public class GetChunkMessage extends ProtocolMessage {
     private final Integer chunk_no;
 
-    public GetChunkMessage(int sender_id, String file_id, Integer chunk_no) {
+    public GetChunkMessage(BigInteger sender_id, String file_id, Integer chunk_no) {
         super(Message_Type.GETCHUNK, sender_id, file_id);
 
         this.chunk_no = chunk_no;
